agents_catalog/32-sila2-lab-automation-agent/bridge_container/requirements.txt
grpcio==1.60.0
grpcio-tools==1.60.0
protobuf==4.25.1
fastapi==0.115.0
uvicorn==0.34.2
pydantic==2.11.0
